+++ NEWS.md
@@ -11,6 +11,7 @@
 * Add Marquez-Megias 2023 adalimumab ([doi:10.3390/biomedicines11102822](https://doi.org/10.3390/biomedicines11102822)) — adults with inflammatory bowel disease.
 * Add Nestorov 2014 factor VIII Fc fusion protein ([doi:10.1002/cpdd.167](https://doi.org/10.1002/cpdd.167)) — previously treated patients with severe hemophilia A.
 * Add Pérez-Ruixo 2025 posdinemab ([doi:10.1002/cpt.70173](https://doi.org/10.1002/cpt.70173)) — healthy adults and adults with Alzheimer's disease.
+* Add Petrov 2024 romiplostim ([doi:10.1002/cpdd.1494](https://doi.org/10.1002/cpdd.1494)) — adults with chronic immune thrombocytopenia (ITP).
 * Add Lin 2024 casirivimab ([doi:10.1007/s11095-024-03764-5](https://doi.org/10.1007/s11095-024-03764-5)) — pediatric and adult subjects, non-infected or with SARS-CoV-2 infection or household contacts.
 * Add Faelens 2021 infliximab ([doi:10.3390/pharmaceutics13101623](https://doi.org/10.3390/pharmaceutics13101623)) — adults with moderate-to-severe ulcerative colitis.
 * Add Sathe 2024 sacituzumab govitecan ([doi:10.1007/s40262-024-01366-3](https://doi.org/10.1007/s40262-024-01366-3)) — adults with metastatic triple-negative breast cancer and other solid tumors.